+++ loader/interfaces.h
@@ -1,7 +1,8 @@
 #include<util.h>
+#include<stdint.h>
 
 //Estas variables no las tengo que liberar si las libero desde la structura
-extern unsigned int interfaceCount;
-extern unsigned int *interface;
+extern uint16_t interfaceCount;
+extern uint16_t *interface;
 
 void parseInterfaces(FILE *r);
